Added NavbarSimple component
This navbar strips down and removes all React Router dependent logic from the Navbar component in order to properly render on Webflow pages that require React Router scripts to be disabled on the front end. NOTE: it is not complete, more of just proof of concept. We should update Webflow before working though anymore styling conflicts